requirements.txt
pandas==2.0.3
numpy==1.25.0
scikit-learn==1.3.0
streamlit==1.25.0
supabase==1.0.3
python-dotenv==1.0.0
joblib==1.3.1
plotly==5.15.0
